fix(authentication): turn ExternalAuthConfig to abstract class; fix circular dependencies
