+++ templates/angular/igx-ts/projects/side-nav-auth/files/src/app/authentication/providers/facebook-provider.ts
@@ -1,8 +1,8 @@
 import { Router } from '@angular/router';
 
 import { AuthProvider } from './auth-provider';
-import { ExternalAuthConfig } from '../services/external-auth.service';
 import { ExternalLogin } from '../models/login';
+import { ExternalAuthConfig } from '../services/external-auth-configs';
 
 export class FacebookProvider implements AuthProvider {
     private user: ExternalLogin;
