+++ templates/angular/igx-ts/projects/side-nav-auth/files/src/app/authentication/services/external-auth-configs.ts
@@ -1,0 +1,18 @@
+export enum ExternalAuthProvider {
+    Facebook = 'Facebook',
+    Google = 'Google',
+    Microsoft = 'Microsoft'
+}
+
+export abstract class ExternalAuthConfig {
+    stsServer: string;
+    client_id: string;
+    scope: string;
+    provider: ExternalAuthProvider;
+    redirect_url: string;
+    response_type: string;
+    post_logout_redirect_uri: string;
+    post_login_route: string;
+    auto_userinfo: boolean;
+    max_id_token_iat_offset_allowed_in_seconds: number;
+}
