+++ templates/angular/igx-ts/projects/side-nav-auth/files/src/app/authentication/services/external-auth.service.ts
@@ -8,30 +8,12 @@ import { FacebookProvider } from '../providers/facebook-provider';
 import { MicrosoftProvider } from '../providers/microsoft-provider';
 import { Location } from '@angular/common';
 import { ExternalLogin } from '../models/login';
-
-export enum ExternalAuthProvider {
-    Facebook = 'Facebook',
-    Google = 'Google',
-    Microsoft = 'Microsoft'
-}
+import { ExternalAuthProvider, ExternalAuthConfig } from './external-auth-configs';
 
 export enum ExternalAuthRedirectUrl {
   Facebook = 'redirect-facebook',
   Google = 'redirect-google',
   Microsoft = 'redirect-microsoft'
-}
-
-export interface ExternalAuthConfig {
-    stsServer: string;
-    client_id: string;
-    scope: string;
-    provider: ExternalAuthProvider;
-    redirect_url: string;
-    response_type: string;
-    post_logout_redirect_uri: string;
-    post_login_route: string;
-    auto_userinfo: boolean;
-    max_id_token_iat_offset_allowed_in_seconds: number;
 }
 
 @Injectable({
